requirements.txt

asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cloudinary==1.22.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-tinymce==3.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
# djangorestframework-simplejwt==4.6.0
drf-jwt==1.17.2
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
packaging==20.4
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sendgrid==6.4.7
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0